requirements.txt
Django
django-filter
djangorestframework
psycopg2